+++ src/main/java/com/supermartijn642/wormhole/extensions/WormholeLevel.java
@@ -1,0 +1,11 @@
+package com.supermartijn642.wormhole.extensions;
+
+import com.supermartijn642.wormhole.PortalGroupCapability;
+
+/**
+ * Created 25/03/2023 by SuperMartijn642
+ */
+public interface WormholeLevel {
+
+    PortalGroupCapability wormholeGetPortalGroupCapability();
+}