+++ src/main/java/com/supermartijn642/wormhole/generator/CoalGeneratorContainer.java
@@ -6,7 +6,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.items.SlotItemHandler;
+import net.neoforged.neoforge.items.SlotItemHandler;
 
 /**
  * Created 12/21/2020 by SuperMartijn642