+++ changelog.md
@@ -1,2 +1,2 @@
 ### Wormhole 1.1.15
-- Initial release of Wormhole for Minecraft 1.20.2
+- Initial release of Wormhole for NeoForge